Created a developer screen that can be worked on by collaborators so that the consumer version isn't affected by the changes.
In this commit, I created seperate pages that follow the according naming conventions so that all of the developers can modify the correct file to apply the changes.

The changes I made can be described by the following: Modified the button on the initial welcome page to the Developer Version and duplicated all of the other pages so that the developers can navigate to the page. I also made buttons to go back to the Consumer UI in order for the developers to compare their current design and then merge it if approved by all of the other members.